reorganizing files a little
