+++ examples/boptest/BOPTest_interface_zone_temp_shed_price.py
@@ -3,16 +3,16 @@ import sys
 sys.path.append("..")
 import os
 
-from controls.hvac.sequences.python.strategies.stra_zone_temp_shed_price import (
+from DFLEXLIBS.hvac.sequences.python.strategies.stra_zone_temp_shed_price import (
     compute_control,
     sparql_query
 )
-from controls.hvac.protocols_zone_temp import (
+from DFLEXLIBS.hvac.protocols_zone_temp import (
     DRControlFunctions,
     DRControlStrategy,
     DRInterface
 )
-from controls.hvac.sequences.python.functions import (
+from DFLEXLIBS.hvac.sequences.python.functions import (
     
     ashrae_TSet_adjust,
     new_comfort_range,
