+++ python/dflexlibs/hvac/functions/fu_shift_occ_price_event.py
