+++ python/dflexlibs/hvac/strategies/__init__.py
@@ -1,2 +1,3 @@
+#TODO: Make sure naming of functions is consistent with the functions directory
 from .stra_zone_temp_shift_shed_price import *
 from .stra_zone_temp_shed_price import *
