+++ python/setup.py
@@ -1,7 +1,7 @@
 from setuptools import setup, find_packages
 
 setup(
-    name="controls",
+    name="dflexlibs",
     version="1.0",
     packages=find_packages(),
 )
